:point_right: Move basic classes to core
