+++ core/index.js
@@ -3,3 +3,5 @@ export * from './entity';
 export * from './context';
 export * from './model-loader';
 export * from './component';
+export * from './tween';
+export * from './material';
