+++ materials/index.js
@@ -1,3 +1,2 @@
-export * from './material';
 export * from './basic';
 export * from './phong';
