+++ tweens/index.js
@@ -1,4 +1,3 @@
-export * from './basic-tween';
 export * from './color-tween';
 export * from './value-tween';
 export * from './vec-tween';
