+++ tweens/vec-tween.js
@@ -1,7 +1,7 @@
 import { vec3 } from '../math';
-import { BasicTween } from './basic-tween';
+import { Tween } from '../core';
 
-export class VecTween extends BasicTween {
+export class VecTween extends Tween {
   action() {
     const _from = [];
     vec3.lerp(_from, this.from, this.to, this.current_step);
